requirements.txt
# Emails.
premailer
cssselect
cssutils
lxml

# Drive integration.
protorpc-standalone==0.9.1
google-api-python-client==1.4.2
oauth2client==1.5.2
httplib2
beautifulsoup4==4.4.0
html2text==2015.6.21

# Testing.
NoseGAE==0.5.6
nose==1.3.1
rednose==0.4.1
coverage==3.7.1
gaenv==0.1.7
PyOpenSSL
